Move `kivy.core.audio` to `kivy.core.audio_output`, reserve `kivy.core.audio_input` for future use #8579 (#8721)
* Renamed kivy.core.audio to kivy.core.audio_output

* Corrected the documentation comment in init.py

* Refactored kivy.core.audio calls in C files

* Reserved audio_input module

* Rename kivy_options from audio to audio_output and changing core_register_libs audio_output call + Change test function calls

* Cleanup of unrelated files

* Fix broken references

---------
