+++ examples/audio/main.py
@@ -17,7 +17,7 @@ kivy.require('1.0.8')
 from kivy.app import App
 from kivy.uix.button import Button
 from kivy.uix.boxlayout import BoxLayout
-from kivy.core.audio import SoundLoader
+from kivy.core.audio_output import SoundLoader
 from kivy.properties import StringProperty, ObjectProperty, NumericProperty
 from glob import glob
 from os.path import dirname, join, basename
