+++ kivy/core/audio_output/audio_android.py
@@ -6,7 +6,7 @@ __all__ = ("SoundAndroidPlayer", )
 
 from jnius import autoclass, java_method, PythonJavaClass
 from android import api_version
-from kivy.core.audio import Sound, SoundLoader
+from kivy.core.audio_output import Sound, SoundLoader
 
 
 MediaPlayer = autoclass("android.media.MediaPlayer")
