+++ setup.py
@@ -979,7 +979,7 @@ if c_options['use_sdl2'] and sdl2_flags:
     for source_file in ('core/window/_window_sdl2.pyx',
                         'core/image/_img_sdl2.pyx',
                         'core/text/_text_sdl2.pyx',
-                        'core/audio/audio_sdl2.pyx',
+                        'core/audio_output/audio_sdl2.pyx',
                         'core/clipboard/_clipboard_sdl2.pyx'):
         sources[source_file] = merge(
             base_flags, sdl2_flags, sdl2_depends)
